Rename directory to wallet-node-client
