+++ wallet/wallet-node-client/src/lib.rs
