+++ wallet/wallet-node-client/src/node_traits.rs
